Pass autosave prop and refactor QuestionnaireResponseForm src-files
