+++ src/containers/PatientDetails/PatientOverviewDynamic/components/StartEncounter/useStartEncounter.ts
@@ -2,7 +2,7 @@ import { Bundle, Encounter } from 'fhir/r4b';
 
 import { extractBundleResources, WithId } from '@beda.software/fhir-react';
 
-import { useQuestionnaireResponseForm } from 'src/components/QuestionnaireResponseForm';
+import { useQuestionnaireResponseForm } from 'src/components/QuestionnaireResponseForm/hooks';
 import { useNavigateToEncounter } from 'src/containers/EncounterDetails/hooks';
 import { StartEncounterProps } from 'src/containers/PatientDetails/PatientOverviewDynamic/components/StartEncounter';
 import { inMemorySaveService } from 'src/hooks/questionnaire-response-form-data';
