+++ src/containers/PatientDetails/index.tsx
@@ -33,11 +33,12 @@ export interface PatientDetailsEmbeddedPageDefinition extends RouteItem {
 export interface PatientDetailsProps {
     embeddedPages?: (patient: Patient, carePlans: CarePlan[]) => PatientDetailsEmbeddedPageDefinition[];
     isDefaultRoutesDisabled?: boolean;
+    autoSave?: boolean;
 }
 
 export const PatientDetails = (props: PatientDetailsProps) => {
     const params = useParams<{ id: string }>();
-    const { isDefaultRoutesDisabled } = props;
+    const { isDefaultRoutesDisabled, autoSave } = props;
 
     const [patientResponse, manager] = usePatientResource({ id: params.id! });
     const author = selectCurrentUserRoleResource();
@@ -109,7 +110,13 @@ export const PatientDetails = (props: PatientDetailsProps) => {
                                             />
                                             <Route
                                                 path="/encounters/:encounterId/new/:questionnaireId"
-                                                element={<PatientDocument patient={patient} author={author} />}
+                                                element={
+                                                    <PatientDocument
+                                                        patient={patient}
+                                                        author={author}
+                                                        autoSave={autoSave}
+                                                    />
+                                                }
                                             />
                                             <Route
                                                 path="/encounters/:encounterId/:qrId/*"
@@ -118,11 +125,19 @@ export const PatientDetails = (props: PatientDetailsProps) => {
                                             <Route path="/documents" element={<PatientDocuments patient={patient} />} />
                                             <Route
                                                 path="/documents/new/:questionnaireId"
-                                                element={<PatientDocument patient={patient} author={author} />}
+                                                element={
+                                                    <PatientDocument
+                                                        patient={patient}
+                                                        author={author}
+                                                        autoSave={autoSave}
+                                                    />
+                                                }
                                             />
                                             <Route
                                                 path="/documents/:qrId/*"
-                                                element={<PatientDocumentDetails patient={patient} />}
+                                                element={
+                                                    <PatientDocumentDetails patient={patient} autoSave={autoSave} />
+                                                }
                                             />
                                             <Route path="/wearables" element={<PatientWearables patient={patient} />} />
                                             <Route
